requirements.txt
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==39.0.1
diffprivlib==0.6.2
flit_core==3.6.0
future==0.18.3
gmpy2==2.1.5
idna==3.4
joblib==1.2.0
mpmath==1.2.1
numpy==1.23.5
phe==1.5.0
Pillow==9.4.0
pip==22.3.1
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.2.1
scipy==1.10.1
setuptools==65.6.3
sympy==1.11.1
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1a0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
wheel==0.38.4
